python/priceparser/requirements.txt
beautifulsoup4
azure-functions
requests